+++ packages/math-input/src/index.ts
@@ -8,6 +8,7 @@ export {CursorContext} from "./components/input/cursor-contexts";
 export {default as KeypadInput} from "./components/input/math-input";
 export {keypadElementPropType} from "./components/prop-types";
 export {default as Keypad} from "./components/provided-keypad";
+export {default as KeypadV2} from "./components/keypad";
 export {default as KeyConfigs} from "./data/key-configs";
 export {default as Keys} from "./data/keys";
 export {KeyType, KeypadType} from "./enums";
